+++ src/Services/Biconnector/BiconnectorServiceBuilder.php
@@ -1,0 +1,40 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Dmitriy Ignatenko <titarx@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Biconnector;
+
+use Bitrix24\SDK\Attributes\ApiServiceBuilderMetadata;
+use Bitrix24\SDK\Core\Credentials\Scope;
+use Bitrix24\SDK\Services\AbstractServiceBuilder;
+use Bitrix24\SDK\Services\Biconnector\Connector\Service\Batch;
+use Bitrix24\SDK\Services\Biconnector\Connector\Service\Connector;
+
+#[ApiServiceBuilderMetadata(new Scope(['biconnector']))]
+class BiconnectorServiceBuilder extends AbstractServiceBuilder
+{
+    /**
+     * Get the Connector service
+     */
+    public function connector(): Connector
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Connector(
+                new Batch($this->batch, $this->log),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+}